packet: Create new interfaces for serializing and deserializing packets
This patch separates 4 different functionalities into 4 different traits

- MqttUncheckedDeserialize: This is for data types that can be
deserialized and have a fixed size, the idea is that for data types with
fixed sizes, one may check for sufficient buffer before hand and
deserialize immediately.

- MqttSized: Tells you the size of the instance of this data type when
serialized, This trait is auto implemented and inlined for types that
implement MqttUncheckedDeserialize since they already have fixed size,
but will have to be manually implemented for other types.

- MqttDeserialize: Deserialize a MQTT data type while performing all
bound checks, this trait is auto implemented for types that implement
MqttUncheckedDeserialize, but has to be implemented manually for other
types.

-MqttSerialize: trait for basic data serialization.

Signed-off-by: cgv6n3qy <apiformes@protonmail.com>

Signed-off-by: cgv6n3qy <apiformes@protonmail.com>