bm:bugfix: Use the right measuring unit for time for subscriber
This bug existed ever since bm was created, but patch 8db563c109b5230b
made it less subtle. The problem was that for some of the publish
functions time was sent in microseconds and for other time was sent in
nano seconds. When refactoring publishers all of them where using
nanosecond at the unit for time measurements, However, the subscribers
were still using milliseconds. This patch addresses this bug by making
ns the measuring unit of time in subscribers.

Signed-off-by: cgv6n3qy <apiformes@protonmail.com>